LayoutTests/http/tests/security/contentSecurityPolicy/worker-blob-inherits-csp-importScripts-block-aborts-all-subsequent-imports-expected.txt

CONSOLE MESSAGE: Refused to load http://localhost:8000/workers/resources/worker-importScripts-source1.js because it does not appear in the script-src directive of the Content Security Policy.
This tests that when calling WorkerGlobalScope.importScripts() will two script URLs that the second script is not loaded if the first script URL was blocked by the Content Security Policy of the worker (inherited from this page).

PASS threw exception Error: NetworkError: DOM Exception 19.
